refactor: move Row class testing to test_row.py
